+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### 2.17.4
 * added `PostsController::getStartAndEndDate()` method;
 * removed `PostsController::indexCompatibility()` method.
+* fixed bug: `PostsTable::getRelated()` returns an empty array if there are no
+    related post;
 * added tests for `PostsCategoriesController` class.
 
 ### 2.17.3